Add, Remove Items functional
Had a weird reload problem connected to the viewModels thing. I need to correct the problems with the old database situation. That is not a happy thing currently. I am not positive how to progress without a clear understanding of what I am doing in the database.

The next step is to connect the bluetooth to the app. We will see how that functions, im sure itll go fine..../s